requirements.txt
joblib==1.0.1
numpy==1.19.2
pycox==0.2.1
tqdm==4.59.0
torch==1.7.0
torchtuples==0.2.0
sklearn_pandas==2.0.3
pandas==1.1.5
deepctr_torch==0.2.6
scikit_learn==0.24.2
